Audio VTS profiling: Push xsd on target for policy config validation
Audio VTS needs to validate /vendor/etc/audio_policy_configuration.xml
with the schema provided with the audio interface.

As the test is run on the target, the xsd needs to be pushed there.

The xsd was correctly pushed for non profiling tests
(thanks to I331d7e4abdf7e520df6d508a8059b6358ad4398f)
but not for profiling tests.
This leads to systematic failure of VtsHalAudioV2_0TargetProfiling
test CheckConfig.audioPolicyConfigurationValidation.

Test: run vts test
Test: vts-tradefed run vts --module VtsHalAudioV2_0ProfilingTarget
Bug: 64444402
Change-Id: Ia5d46ad893600d0aea1b51dfdf72a4fdca71f9b4
Signed-off-by: Kevin Rocard <krocard@google.com>