Audio VTS: Enable batch mode
Bug: 36819162

Batch mode speeds up tests from 14min to 1min.
The only drawback is that hal deinit will not be as well tested as
before. HAL init/deinit used to be called 1k+ times but in batch
mode it will only be called once.

This is not so much of a problem as hal are not designed to start/stop
multiple times anyway.

Test: vts-tradefed run commandAndExit vts  --module VtsHalAudioV2_0Target
Test: watch -d 'adb shell ps |grep HalAudio'
Change-Id: I84aef8bc08f34f14fac9a26b65b7f6b118a77327
Signed-off-by: Kevin Rocard <krocard@google.com>