vts_treble_vintf_test: (is served) => (in manifest)
This is a relationship we had previously thought was satisfied
(because it's enforced by getService), but client code is avoiding this
check in some cases.

Take for instance the recommended pattern for working with multiple
minor versions:
mFoo0 = V1_0::IFoo::getService();
mFoo1 = V1_1::IFoo::castFrom(mFoo0);

This means that only 1.0 being in the manifest is being enforced
(and so 1.1 being served isn't actually a requirement in the manifest).

Test: vts_treble_vintf_test --gtest_filter=VtsTrebleVintfTest.ServedHalsAreInManifest
    passes on walleye.
Bug: 73772212

Change-Id: Ie5fcc991a8b47eeff94baa59fc7146f7a3e0be29