VtsTrebleVintfTest: provide all instances to CheckDeprecation.
It is possible that matrix specify a regular expression on instances,
so there is no way to find an exact instance name and use GetHalService.
Use listByInterface() instead to get all instances and report
to CheckDeprecation.

For passthrough service manager, listByInterface doesn't work, so
a list of instance names are provided by CheckDeprecation as a hint.
This list includes the legacy list of <instance> tags. Hence,
the deprecation of <regex-instance> instances cannot be enforced.

Bug: 73738616
Test: run this test on walleye
Change-Id: I210dfebba3155926fe5737d3a318acb5b06b6583