vts_treble_vintf_test: <= O-MR1, less strict
In P, several new requirements were added universally across
devices. However, because old architectures were too dangerous
to change, we're opting to only add these requirements starting
on devices which launch with P. For instance, certain HAL
implementations were tightly coupled with the hardware they talk
to and could only start up when the device is in a specific state.
This means that devices launching with O/O-MR1 in P will not
be guaranteed to have an accurate vendor manifest. This is not
a regression from when they were launched.

For similar reasons, the hashes for interfaces in these old
architectures can not be updated, so they have also been exempted.

If a device has warnings, this means a side-band ABI is being
exposed which will need to be maintained in a backwards
compatible way.

Bug: 68662126 # for test failures
Bug: 77981855 # for getService not returning
Bug: 68661741 # for curren.txt issues
Test: vts_treble_vintf_test_all passes on Pixel 2
Test: vts_treble_vintf_test_all no change on device launching with P

Change-Id: Id7ddc88ed29589c203ea78a5f3cc30e3852c4727