Revert "Add android.hardware.tetheroffload.config/control HALs to the exempt list."
This reverts commit a199a00600f67eeb5f9ff0fd2a3e4afbe1035ed8.

Bug: 69285625

Reason for revert: this is the wrong exemption list. Tether offload was added here in order to fix failures in in interfaces are released, but the correct way to do that would have been to backport the hash. It is still required for to be served if it is available, and this change did not fix the original issue.

Change-Id: If64777883d25bcdfc271ffb317d97a87573c92f1