VtsTrebleVintfVendorTest: GNSS test for >Q.
This test explicitly failed on API levels greater than Q in case the
VINTF FCM has a better way of expressing this test, to give us an
opportunity to remove the test.

Actually, this test should remain indefinitely. Likely, only additional
cases of this would encourge an overhaul of the VINTF FCM format for HAL
requirements. If we do have that, there is no harm in keeping the test,
and if we do that, we should be able to discover this test anyway.

Bug: 147387997
Test: VtsTrebleVintfVendorTest
Change-Id: I17b1349aed3d0d94b4753aa888cf8079d6f8398f