shipping level 31, target-level 6
Updating this test to reflect "S" devices. WARNING: THIS IS
COMPLETELY PROVISIONAL. The shipping API and target levels for Android S
have not been finalized (or even R at this stage). These API values are
only meaningful in release Android branches, and they are subject to
change here at anytime. They exist here only as an integration test for
the platform.

Bug: 150215458
Test: atest vts_treble_vintf_vendor_test
Change-Id: Ibf04618931e86414d7355c8316b1040535811114