Fix vts_treble_vintf_vendor_test for updatable HALs
DeviceManifest/SingleManifestTest#ManifestAidlHalsServed was checking
that every AIDL HAL is served with the exact version declared in the
manifests.

This can fail on a device with a HAL which is updated via APEX. At
runtime this is not a problem because the new version of the interface
can talk to clients expecting older interface.

Now, the test allows version upgrades for those updatable HALs if the
delcared APEX is actually updated.

Bug: 130058564
Bug: 175082690
Test: atest vts_treble_vintf_vendor_test
  with updated HAL instance with "updatable-via-apex"
  should fail with prebuilt apexes (no updates)
  should pass with installed apexes (updates)
Change-Id: I9a626c191e09f8e04197afb7050c6dfef1b91969