vts_treble_vintf_vendor_test: AIDL R ver exempt
VINTF manifest versioning for Android R did not exist. Since the only
AOSP HAL interface in Android R which had a version (identity) was
already exempted, this doesn't hurt the worst case deprecation
requirements for interfaces released at that time. However, in the
opportunistic case, where all old devices elect to upgrade to a newer
HAL version, we may not know this, and miss deprecation window for a
year. In any case, this is a small and unlikely cost to pay, and given
GRF, we are unlikely to find the benefit anyway.

One alternative to consider here would be to have a separate requirement
for GRF devices and non-GRF devices. However, I think this would be too
much complexity - keeping the requirements the same makes it easier to
test the requirements, and it makes existing R devices easier to upgrade
to the next release if they are non-GRF.

Bug: 199190514
Test: vts_treble_vintf_vendor_test
Change-Id: I990aeb71e0aa61e3f6a0170761c828cf75bc5d59
Merged-In: I990aeb71e0aa61e3f6a0170761c828cf75bc5d59