Treble VINTF test: parameterize on HAL instances.
Before the change, some tests use a for loop to
iteratively query each HAL for information. When
a subset of the HAL fails, the whole test fails.

This change parameterize these tests on the
HAL name so each HAL instance has its own test case.
This allows faster triaging and declared
clearer ownership when handling test failures.

Note: As a side-effect of this change, time outs are only
imposed on certain binder transactions (e.g.
getting a HAL service), but no longer
on the whole test case of a HAL to avoid over complicating
the test. Hence,
it is technically possible that the operations after
getting the HAL service (e.g. getting hash chain,
getting debug info, etc.) are stuck. However, the usual
case is that a HAL is not responsive before the test,
which will hit the timeout at getting the HAL service
as the first step. In addition, even without this change,
when the whole test case times out, a warning message
is printed but no error is raised.

Bug: 240314930
Test: vts_treble_vintf_test_all
Test: atest vts_treble_vintf_vendor_test
Test: atest vts_treble_vintf_framework_test

Change-Id: I657b86cfcc266a83c10dbfd90692eb7286cc5271