Make sure we don't have undeclared HALs in the device manifest
This is an error that is typically caught at build-time. It's possible
for a device to work around the build time failure and it may cause
unexpected problems.
So now we check in VTS to make sure if a VINTF HAL is declared in the
device manifest, it is also declared in the Framework Compatibilty
Matrix.

Test: atest vts_treble_vintf_vendor_test
Bug: 379673733
Change-Id: I77e93cc1b0349aace8ee0de145677ce8214d4e98