vts_treble_vintf_vendor_test: check AIDL partition
We had this test in HIDL. However, for AIDL, I did not port
the test during the Unify Android IDL project because it
was not a common issue and it never came up.

However, due to the increasing complexity of Rust implementations,
as well as the use of APEX (for which this doesn't work yet),
going ahead and adding the test.

This is important because HALs added to the vendor manifest which
are really on system don't actually have to be supported for
as many years, but if they are configured wrong today, we would
have to support them.

Bug: 385355208
Test: atest vts_treble_vintf_vendor_test
Change-Id: I2142d6b6bd2518360d873b61c3a61c80f50b61ff