vts_treble_vintf_trusted_hal_test: use nohup for the "vm run" command
Make the test_vm setup resilient to adb connection resets.
With the recent AVF improvement where the crosvm process is killed
when its parent's virtmgr process is killed, a VM started via
the utility "vm run" is now killed when the vm run's
adb session is killed.
This makes the test_vm setup very brittle and sensible to
adb session flakes, which is what is observed when
using Tradefed's RunCommandTargetPreparer background command.
To solve this, nohup is used to prevent the vm run process
to be killed upon an adb session reset.
A new teardown command is also added to kill the process started
by nohup when the test completes.

Bug: 420649756
Flag: EXEMPT bug fix
Test: atest vts_treble_vintf_trusted_hal_test
Change-Id: I83c62c3ee704bb3c21fc7087b8c1cf37f855293d