trusted_hal_vintf: Make test names stable across environments
Prior to this cl, The gtest parameterizer appended an index to
test names (e.g., `..._6`). This index depends on the order and number
of HALs discovered on the target device during testing.

On a local device with a full set of HALs (including Widevine DRM), a
test for ISecureStorage might be named `..._7`. On a CI environment
where some HALs are not present, the same test might be named `..._6`.

This name instability makes it impossible to reliably skip or manage
these tests using configuration files, as the test's full name is not
predictable.

This commit updates `GetTestCaseSuffix` to conditionally omit the index
from the test name suffix when the `TRUSTED_HAL_TEST` macro is defined.
This change ensures that Trusted HAL test names are consistent and
stable, regardless of the specific HALs present on the device.

Bug: 432948133
Test: atest vts_treble_vintf_trusted_hal_test \
vts_treble_vintf_vendor_test
Flag: EXEMPT trusted hal test name improvement

Change-Id: If79c132b33955ee86a8611bc6dc36c816a7f78eb