Check vm-exclusive HALs for default libbinder stability
Trusted HALs are not running in Android, but rather a TEE. This makes it
difficult to validate that they are being served with a libbinder built
with the correct default stability settings. This is important because
mistakenly serving a vendor HAL with a system-stability libbinder could
lead to defacto stabilization of interfaces that AOSP or OEMs did not
mean to stabilize of freeze.

This tests checks that the default stability of each vm-exclusive HAL is VENDOR. It relies on each Trusted HAL adding a non-vintf extension to their service.

Bug: 462787437
Test: atest vts_treble_vintf_trusted_hal_test
Change-Id: If1a28d57755ed41a5b961897106b9b806e7f273d